♻️ refactor(server): decompose Application.kt into focused sibling files
